backend-2/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
lightgbm
pandas
joblib
google-genai
python-dotenv
